Fix checkstyle issues (#7058)
