+++ realm/realm-library/src/main/java/io/realm/rx/CollectionChange.java
