+++ realm/realm-library/src/objectServer/java/io/realm/mongodb/App.java
@@ -156,7 +156,7 @@ public class App {
     // Currently we only allow one instance of App (due to restrictions in ObjectStore that
     // only allows one underlying SyncClient).
     // FIXME: Lift this restriction so it is possible to create multiple app instances.
-    public volatile static boolean CREATED = false;
+    public static volatile boolean CREATED = false;
 
     /**
      * Thread pool used when doing network requests against MongoDB Realm.
